Repackaged 5.1.49 with added 'mongo' package name.
In order to upgrade Cluvio's MySQL JDBC driver without breaking
backward-compatibility with users of the MongoDB BI Connector, which
only works with version 5.1. of the MySQL JDBC driver, we re-package
the MySQL JDBC driver with a new package name, so that we can use it
for connections to Mongo BI connectors while using a newer MySQL JDBC
driver for native MySQL connections.